+++ parler/fields.py
@@ -94,7 +94,9 @@ class TranslatedFieldDescriptor(object):
         return "<{0} for {1}.{2}>".format(self.__class__.__name__, self.field.model.__name__, self.field.name)
 
     def short_description(self):
-        """ Renders the correct verbose name for the translated fields"""
+        """
+        Ensure that the admin ``list_display`` renders the correct verbose name for translated fields.
+        """
         return self.field.model._translations_model._meta.get_field_by_name(self.field.name)[0].verbose_name